devops_integration/python_api/requirements.txt
jsonpickle
cloudshell-automation-api>=7.1,<7.2